:recycle: refactor: 숙소 상세 응답 dto 패키지 이동
